Table For Now
This is turning out to be very complicated. My internal dynamic rate advance after the fact and the initial phi both make it non-representative of the actual value of omega at any instant of time. I need another way to determine the current omegaN and such. To make matters worse, I really need to document what happens when it accelerates past the nyquist frequency. What then?